Fix crash bugs and improve sticky windows behavior
Crash fixes:
- Fix segfault in CLV_AddItem when adding files before playlist window exists
  - Add NULL check for globals.m_hPlaylistViewControl in CPL_cb_OnPlaylistAppend
  - Add NULL check in CPL_cb_OnPlaylistItemDelete
  - Add NULL checks in CPL_HandleAsyncNotify for CLV_BeginBatch/EndBatch

- Fix segfault when toggling playlist window visibility
  - Add NULL check in CPlaylistWindow_SetVisible

- Fix segfault on application exit
  - Add NULL check in CPlaylistWindow_Destroy before IF_CloseWindow
  - Set windows.m_hifPlaylist to NULL after closing

- Fix corrupted code in main.c where CPlaylistWindow_Create() was
  embedded as a literal string with escaped newlines, preventing the
  playlist window from ever being created

Sticky windows improvements:
- Only snap when windows are actually edge-to-edge (adjacent), not
  just when edges are aligned at any distance
- Add UNDOCK_DISTANCE (12px) - drag threshold to break docking
- Add justUndocked flag to prevent immediate re-snapping after undocking
- Reduce SNAP_DISTANCE from 12 to 6 pixels for less aggressive snapping

Options dialog fix:
- Fix sticky_windows option not being saved when clicking OK
  - Add missing BM_GETCHECK call for IDC_STICKYWINDOWS

Build warning fixes:
- Add (void) casts to 8 path_remove_filespec calls in skin.c to
  suppress warn_unused_result warnings

+++ profile.c
@@ -238,7 +238,7 @@ void    options_read(void)
 	options.equalizer =
 		GetPrivateProfileInt("Equalizer", "Active", 0, pathbuf);
 	    
-	for (teller = 1; teller < ARRAY_SIZE(options.eq_settings); teller++)
+	for (teller = 1; teller < (int)ARRAY_SIZE(options.eq_settings); teller++)
 	{
 		char    keyname[100];
 		snprintf(keyname, sizeof(keyname), "Eq%d", teller);
@@ -502,7 +502,7 @@ void    options_write(void)
 							  intbuf,
 							  pathbuf);
 	
-	for (teller = 1; teller < ARRAY_SIZE(options.eq_settings); teller++)
+	for (teller = 1; teller < (int)ARRAY_SIZE(options.eq_settings); teller++)
 	{
 		char    keyname[100];
 		snprintf(keyname, sizeof(keyname), "Eq%d", teller);
